Loosen overly-tight restrictions from $hookctl(SET .. SERIAL) and $hookctl(SET .. NICK)
It's allowed to have two hooks on the same list with the same serial as long as they have
different patterns, and two hooks on the same list with the same pattern as long as they
have different serials.

The previous code for $hookctl(SET .. SERIAL) wouldn't allow you to set a serial if there
was any other hook on the same list with the same serial.

Similarly, the previous code for $hookctl(SET .. NICK) wouldn't allow you to set a pattern
if there was any other hook on the same list with the same pattern.