Fix $split() in the case of inverted, multiple-character arguments
$split(^abc aXXbXXc) should return "a  b  c".

It isn't the individual found/not-found checks for each character in "abc"
that need to be inverted: it's the overall "is the character in the set" check
that should be inverted.

ie. it should be NOT (a OR b OR c), rather than (NOT a OR NOT b OR NOT c).